refactor(PT-70): move Footer to shared and adjust spacing between sections
