+++ src/shared/components/Layout/Footer/Footer.tsx
@@ -1,6 +1,6 @@
-import { footerSections } from '@/features/Footer/footerData';
+import { footerSections } from '@/shared/components/Layout/Footer/footerData';
 import { Link } from 'react-router-dom';
-import Accordion from '@/features/Footer/Accordion';
+import Accordion from '@/shared/components/Layout/Footer/Accordion';
 import Logo from '@/shared/components/UI/Logo';
 
 const Footer = () => {
@@ -8,21 +8,21 @@ const Footer = () => {
     <footer
       role="contentinfo"
       className="w-full mx-auto text-alabaster bg-gradient-to-t from-red-beech to-fiery-tenn rounded-t-[60px]
-        px-[26px] pt-[44px] pb-[27px] max-w-[375px]
+        px-10 py-10 max-w-[375px]
         xl:px-[120px] xl:pt-[29px] xl:pb-[25px] xl:max-w-[1280px]"
     >
       {/* Логотип */}
-      <div className="flex justify-center  mb-[73px] xl:mb-[39px]">
+      <div className="flex justify-center mb-[73px] xl:mb-[39px]">
       <Logo
       textColor="text-alabaster"
       iconFill="fill-alabaster"/>
       </div>
       
       {/* Mobile: Accordion */}
-      <div className="flex flex-col text-lg gap-7 mb-[61px] xl:hidden">
+      <div className="flex flex-col text-lg gap-5 mb-[61px] xl:hidden">
         {footerSections.map(({ title, links }) => (
           <Accordion key={title} title={title}>
-            <ul className="flex flex-col text-sm gap-4">
+            <ul className="flex flex-col mt-[20px] text-sm gap-4">
               {links.map(({ text, to }) => (
                 <li key={to}>
                   <Link
