+++ src/shared/components/Layout/Footer/footerData.ts
