+++ src/shared/components/Layout/MainLayout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Footer from '@/features/Footer/Footer';
+import Footer from '@/shared/components/Layout/Footer/Footer';
 import Header from '@/components/Header/Header';
 import { Suspense } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
